atoi

+++ ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: avarnier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/14 14:45:28 by avarnier          #+#    #+#             */
-/*   Updated: 2019/12/01 10:46:20 by avarnier         ###   ########.fr       */
+/*   Updated: 2019/12/01 10:52:49 by avarnier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,10 +32,10 @@ static size_t	ft_start(const char *s1, const char *set)
 	return (x);
 }
 
-static int	ft_end(const char *s1, const char *set)
+static size_t	ft_end(const char *s1, const char *set)
 {
-	int x;
-	int y;
+	size_t	x;
+	size_t	y;
 
 	x = ft_strlen(s1) - 1;
 	y = 0;
@@ -52,7 +52,7 @@ static int	ft_end(const char *s1, const char *set)
 	return (x);
 }
 
-char		*ft_strtrim(const char *s1, const char *set)
+char			*ft_strtrim(const char *s1, const char *set)
 {
 	if (ft_start(s1, set) == ft_strlen(s1))
 		return (ft_strdup(""));
